Allow to add baseline in Submodules
