--- detekt-gradle-plugin/src/test/kotlin/io/gitlab/arturbosch/detekt/testkit/Submodule.kt
@@ -1,9 +1,0 @@
-package io.gitlab.arturbosch.detekt.testkit
-
-data class Submodule(
-    val name: String,
-    val numberOfSourceFilesPerSourceDir: Int,
-    val numberOfCodeSmells: Int,
-    val buildFileContent: String?,
-    val srcDirs: List<String>
-)
